react native alert

+++ ReactNative/confusion/components/FavoriteComponent.js
@@ -1,8 +1,9 @@
 import * as React from 'react';
-import { View, Text, FlatList } from 'react-native';
+import { View, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
+import Loading from './LoadingComponent';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
 
@@ -28,7 +29,23 @@ class Favorites extends React.Component {
         {
           text: 'Delete',
           type: 'delete',
-          onPress: () => this.props.deleteFavorite(item.id)
+          onPress: () => {
+            Alert.alert('Delete favorite',
+                'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                [
+                  {
+                    text: 'Cancel',
+                    style: 'cancel',
+                    onPress: () => console.log(item.name + ' Not Deleted')
+                  },
+                  {
+                    text: 'OK',
+                    onPress: () => this.props.deleteFavorite(item.id)
+                  }
+                ],
+                { cancelable: false }
+            );
+          }
         }
       ];
